rename protocol to common. move physics, rendering there, to make it accessible to both worker and main thread
