+++ src/common/growableTypedArray.ts
