+++ src/common/protocol.ts
