+++ src/common/tsconfig.json
@@ -17,6 +17,8 @@
     "moduleResolution": "node",
     "composite": true,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "isolatedModules": true,
+    "esModuleInterop": true
   }
 }
