+++ src/worker/index.ts
@@ -1,74 +1,20 @@
-import type { MutateMatrix, GetDrawBuffer, MainLoop, ShouldRun, GetPixelsPerMeter } from './onContext'
-import { onContext } from './onContext'
-import { Demo } from '../protocol'
-import type { FromMain, ReadyFromWorker } from '../protocol'
-import { DrawBuffer, drawBuffer, flushDrawBuffer } from './debugDraw'
-import type { DestroyDemo, WorldStep } from './demo'
-import type { mat3 } from 'gl-matrix'
+import { onContext } from '../common/onContext'
+import type { FromMain, ReadyFromWorker } from '../common/protocol'
+import {
+  shouldRun,
+  mainLoop,
+  getDrawBuffer,
+  flushDrawBuffer,
+  mutateMatrix,
+  pixelsPerMeterGetter,
+  switchDemo,
+  setClearCanvas
+} from '../common/demoSwitcher'
 
 self.onmessageerror = (event: MessageEvent) =>
   console.error('onmessageerror', event)
 self.onerror = (event: ErrorEvent) =>
   console.error('onerror', event)
-
-type ClearCanvas = () => void
-
-let world: Box2D.b2World | undefined
-let destroyDemo: DestroyDemo | undefined
-let worldStep: WorldStep | undefined
-let clearCanvas: ClearCanvas | undefined
-let matrixMutator: MutateMatrix | undefined
-let getPixelsPerMeter: GetPixelsPerMeter | undefined
-
-const { debugDraw } = await import('./debugDraw')
-
-const frameLimit = 60
-const frameIntervalMs = 1 / frameLimit * 1000
-const minDrawIntervalMs = frameIntervalMs / 2
-
-const switchDemo = async (proposedDemo: Demo): Promise<void> => {
-  destroyDemo?.()
-  destroyDemo = undefined
-  worldStep = undefined
-  matrixMutator = undefined
-  getPixelsPerMeter = undefined
-  clearCanvas?.()
-  switch (proposedDemo) {
-    case Demo.None:
-      world = undefined
-      break
-    case Demo.Ramp: {
-      const boxCount = 100
-      const { makeRampDemo } = await import('./demo/ramp');
-      ({ world, destroyDemo, worldStep, matrixMutator, getPixelsPerMeter } = makeRampDemo(debugDraw, boxCount))
-      break
-    }
-    case Demo.WaveMachine: {
-      const { makeWaveMachineDemo } = await import('./demo/waveMachine');
-      ({ world, destroyDemo, worldStep, matrixMutator, getPixelsPerMeter } = makeWaveMachineDemo(debugDraw, frameLimit))
-      break
-    }
-    default:
-      throw new Error(`Unsupported demo type '${proposedDemo as string}'`)
-  }
-}
-
-const shouldRun: ShouldRun = (intervalMs: number): boolean =>
-  intervalMs > minDrawIntervalMs && world !== undefined
-
-const mainLoop: MainLoop = (intervalMs: number): void =>
-  worldStep?.(intervalMs)
-
-const getDrawBuffer: GetDrawBuffer = (): DrawBuffer => {
-  world?.DebugDraw()
-  return drawBuffer
-}
-
-const mutateMatrix: MutateMatrix = (out: mat3, canvasWidth: number, canvasHeight: number): void =>
-  matrixMutator?.(out, canvasWidth, canvasHeight)
-
-const pixelsPerMeterGetter: GetPixelsPerMeter = (): number =>
-  getPixelsPerMeter?.() ?? 32
 
 self.onmessage = ({ data }: MessageEvent<FromMain>) => {
   switch (data.type) {
@@ -77,7 +23,7 @@ self.onmessage = ({ data }: MessageEvent<FromMain>) => {
       if (gl === null) {
         throw new Error('Failed to create WebGL2 rendering context')
       }
-      clearCanvas = () => gl.clear(gl.COLOR_BUFFER_BIT)
+      setClearCanvas(() => gl.clear(gl.COLOR_BUFFER_BIT))
       onContext(
         gl,
         shouldRun,
